requirements.txt
openai>=1.0.0
paho-mqtt
pyyaml
